casting longlong to int to pacify compiler.
In the long run, we want to change that type into something bigger. However, currently, the backend uses a N_num to encode this which, internally uses int....
We cannot even make it unsigned as we use -1 to indicate lack of ravel knowledge!

+++ src/libsac2c/typecheck/type_utils.c
@@ -1823,7 +1823,7 @@ int TUgetLengthEncoding (ntype *type)
     } else if (TUisScalar (type)) {
         res = 0;
     } else {
-        res = SHgetUnrLen (TYgetShape (type));
+        res = (int)SHgetUnrLen (TYgetShape (type));
     }
 
     DBUG_RETURN (res);
